adapt the tests
